moving get_input_column_types() to Connector
Changes to be committed:
	modified:   src/predictions/rudderstack_predictions/connectors/Connector.py
	modified:   src/predictions/rudderstack_predictions/train.py
	modified:   src/predictions/rudderstack_predictions/wht/pythonWHT.py
